requirements.txt
python-binance==0.7.4
pandas
numpy
python==3.7